fix: dots.txt
create dots.txt if doesnt exist, verify if that file exists before doing
any action

+++ dots.sh
@@ -2,9 +2,17 @@
 export DOTS_DIR="${XDG_DATA_HOME:-$HOME}/dotfiles"
 export DOTS_DATA_FILE="$DOTS_DIR/dots.txt"
 
+if [ ! -e "$DOTS_DATA_FILE" ]; then
+  touch "$DOTS_DATA_FILE"
+fi
+
 function _dots_remove() {
-  declare config answer
-  config=$(gum choose $(grep "" "$DOTS_DATA_FILE"))
+  declare config answer data
+  data=$(grep "" "$DOTS_DATA_FILE")
+
+  if [ -n "$data" ]; then
+    config=$(gum choose $data)
+  fi
 
   if [ -n "$config" ] && [ -e "$HOME/$config" ]; then
     gum style --foreground 37 "the followings commands are going to be executed:"
@@ -12,8 +20,9 @@ function _dots_remove() {
     gum confirm "proceed with these commands?" && answer="y"
 
     if [ "$answer" = "y" ]; then
-      gum spin --spinner dot --title "removing $HOME/$config" -- sleep 3
+      gum spin --spinner dot --title "preparing to remove $HOME/$config" -- sleep 3
       rm -rf "$HOME/$config"
+      gum spin --spinner dot --title "removing $HOME/$config" -- sleep 2
       gum log --structured --level info "the $HOME/$config has been removed"
     else
       gum log --structured --level info "operation canceled"
@@ -27,7 +36,7 @@ function _dots_list() {
   if [ ! -e "$DOTS_DATA_FILE" ]; then
     touch "$DOTS_DATA_FILE"
   fi
-  gum pager <"$DOTS_DATA_FILE"
+  less "$DOTS_DATA_FILE"
 }
 
 function _dots_installation() {
@@ -39,8 +48,9 @@ function _dots_installation() {
   gum confirm "proceed with these commands?" && answer="y"
 
   if [ "$answer" = y ]; then
-    gum spin --spinner dot --title "installing $DOTS_DIR/$config to $HOME/$config" -- sleep 3
+    gum spin --spinner dot --title "preparing to install $DOTS_DIR/$config to $HOME/$config" -- sleep 3
     cp -r "$DOTS_DIR/$config" "$HOME/$config"
+    gum spin --spinner dot --title "installing $DOTS_DIR/$config to $HOME/$config" -- sleep 2
     gum log --structured --level info "the $config has been installed"
   else
     gum log --structured --level info "operation canceled"
@@ -52,14 +62,18 @@ function _dots_add() {
 }
 
 function _dots_install() {
-  declare config answer
-  config=$(gum choose $(grep "" "$DOTS_DATA_FILE"))
+  declare config answer data
+  data=$(grep "" "$DOTS_DATA_FILE")
+
+  if [ -n "$data" ]; then
+    config=$(gum choose $data)
+  fi
 
   if [ "$config" == ".bashrc" ]; then
+    gum spin --spinner dot --title "preparing to install $config..." -- sleep 3
     echo "source $DOTS_DIR/$config" >>"$HOME/.bashrc"
-    gum spin --spinner dot --title "installing $config..." -- sleep 3
-    gum spin --spinner dot --title "exiting..." -- sleep 1
-    gum style --foreground 64  "completd"
+    gum spin --spinner dot --title "installing $config..." -- sleep 2
+    gum style --foreground 64 "completd"
     return
   fi
 
@@ -69,8 +83,9 @@ function _dots_install() {
       gum confirm "Do you want to overwrite it?" && answer="y"
 
       if [ "$answer" = "y" ]; then
-        gum spin --spinner dot --title "removing '$HOME/$config..." -- sleep 3
+        gum spin --spinner dot --title "preparing to remove '$HOME/$config..." -- sleep 3
         rm -rf "$HOME/$config"
+        gum spin --spinner dot --title "removing '$HOME/$config..." -- sleep 2
         _dots_installation "$config"
       else
         gum log --structured --level info "operation canceled: $answer"
@@ -84,21 +99,27 @@ function _dots_install() {
 }
 
 function _dots_update() {
-  declare config answer
-  config=$(gum choose $(grep "" "$DOTS_DATA_FILE"))
+  declare config answer data
 
-  if [ -n "$config" ] || [ -e "$HOME/$config" ]; then
+  data=$(grep "" "$DOTS_DATA_FILE")
+
+  if [ -n "$data" ]; then
+    config=$(gum choose $data)
+  fi
+
+  if [ -n "$config" ] && [ -e "$HOME/$config" ]; then
     gum style --foreground 37 "the followings commands are going to be executed:"
     gum style --foreground 160 "rm -rf '$DOTS_DIR/$config'"
     gum style --foreground 160 "cp -r '$HOME/$config' '$DOTS_DIR/$config'"
     gum confirm "proceed with these commands?" && answer="y"
 
     if [ "$answer" = "y" ]; then
-      gum spin --spinner dot --title "removing '$DOTS_DIR/$config'..." -- sleep 5
+      gum spin --spinner dot --title "preparing to remove '$DOTS_DIR/$config'..." -- sleep 3
       rm -rf "$DOTS_DIR/$config"
-      gum style --foreground 37 "copying '$HOME/$config' to '$DOTS_DIR/$config'"
-      gum spin --spinner dot --title "copying '$HOME/$config' to '$DOTS_DIR/$config'..." -- sleep 5
+      gum spin --spinner dot --title "removing '$DOTS_DIR/$config'..." -- sleep 2
+      gum spin --spinner dot --title "preparing to copy '$HOME/$config' to '$DOTS_DIR/$config'..." -- sleep 3
       cp -r "$HOME/$config" "$DOTS_DIR/$config"
+      gum spin --spinner dot --title "copying '$HOME/$config' to '$DOTS_DIR/$config'..." -- sleep 2
       gum log --structured --level info "the $DOTS_DIR/$config  has been updated"
     else
       gum log --structured --level info "operation canceled"
